backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
boto3>=1.34.0
botocore>=1.34.0
amazon-transcribe>=2.0.1
moto[dynamodb]>=4.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
